Create a separate folder for multidevice tests
