+++ tests/python/multidevice/test_communication.py
@@ -5,12 +5,12 @@
 import pytest
 import torch
 
-import multidevice_fixtures
+import fixtures
 import nvfuser
 from nvfuser import DataType, FusionDefinition
 
 
-multidevice_test = multidevice_fixtures.multidevice_test
+multidevice_test = fixtures.multidevice_test
 
 
 @pytest.mark.mpi
