+++ tests/python/multidevice/test_multidevice.py
@@ -7,12 +7,12 @@ import torch
 from enum import Enum, auto
 from torch.nn.attention import SDPBackend
 
-import multidevice_fixtures
+import fixtures
 import nvfuser
 from nvfuser import DataType, FusionDefinition
 from nvfuser.testing.utils import create_sdpa_rng_tensors, define_sdpa_rng_state
 
-multidevice_test = multidevice_fixtures.multidevice_test
+multidevice_test = fixtures.multidevice_test
 
 
 @pytest.mark.mpi
